+++ foliant/preprocessors/plantuml.py
@@ -3,7 +3,7 @@ PlantUML diagrams preprocessor for Foliant documenation authoring tool.
 '''
 
 import re
-from pathlib import Path
+from pathlib import Path, PosixPath
 from hashlib import md5
 from subprocess import run, PIPE, STDOUT, CalledProcessError
 from typing import Dict
@@ -11,6 +11,7 @@ OptionValue = int or float or bool or str
 
 from foliant.preprocessors.base import BasePreprocessor
 from foliant.utils import output
+from foliant.preprocessors.utils.combined_options import CombinedOptions
 
 
 class Preprocessor(BasePreprocessor):
@@ -18,6 +19,8 @@ class Preprocessor(BasePreprocessor):
         'cache_dir': Path('.diagramscache'),
         'plantuml_path': 'plantuml',
         'parse_raw': False,
+        'format': 'png',
+        'as_image': True
     }
     tags = 'plantuml',
 
@@ -33,7 +36,7 @@ class Preprocessor(BasePreprocessor):
     def _get_command(
             self,
             options: Dict[str, OptionValue],
-            diagram_src_path: Path
+            diagram_src_path: PosixPath
         ) -> str:
         '''Generate the image generation command. Options from the config definition are passed
         as command options (``cache_dir`` and ``plantuml_path`` options are omitted).
@@ -44,19 +47,23 @@ class Preprocessor(BasePreprocessor):
         :returns: Complete image generation command
         '''
 
-        components = [self.options['plantuml_path']]
+        components = [options['plantuml_path']]
 
-        params = self.options.get('params', {})
+        params = dict(options.get('params', {}))
 
-        for option_name, option_value in {**params, **options}.items():
-            if option_name == 'caption':
-                continue
+        # if format is not set in parameters, set it from `format` param
+        if 't' not in (k[0] for k in params.keys()):
+            params[f't{options["format"]}'] = True
 
-            elif option_value is True:
+        for option_name, option_value in params.items():
+            # if option_name == 'caption':
+            #     continue
+
+            if option_value is True:
                 components.append(f'-{option_name}')
 
-            elif option_name == 'format':
-                components.append(f'-t{option_value}')
+            # elif option_name == 'format':
+            #     components.append(f'-t{option_value}')
 
             else:
                 components.append(f'-{option_name} {option_value}')
@@ -65,7 +72,35 @@ class Preprocessor(BasePreprocessor):
 
         return ' '.join(components)
 
-    def _process_plantuml(self, options: Dict[str, OptionValue], body: str) -> str:
+    def _get_result(self, diagram_path: PosixPath, options: CombinedOptions):
+        '''Get either image ref or raw image code depending on as_image option'''
+        diagram_format = self._get_diagram_format(options)
+        if diagram_format != 'svg' or options['as_image']:
+            result = f'![{options.get("caption", "")}]({diagram_path.absolute().as_posix()})'
+        else:
+            with open(diagram_path, 'r') as f:
+                result = f'<div>{f.read()}</div>'
+                # return plantuml md5 comment because it contains diagram definition
+                md5_pattern = re.compile(r'<!--MD5.+?-->', re.DOTALL)
+                result = md5_pattern.sub('', result)
+        return result
+
+    def _get_diagram_format(self, options: CombinedOptions) -> str:
+        '''
+        Parse options and get the final diagram format. Format stated in params
+        (e.g. tsvg) has higher priority.
+
+        :param options: the options object to be parsed
+
+        :returns: the diagram format string
+        '''
+        result = None
+        for key in options.get('params', {}):
+            if key.startswith('t'):
+                result = key[1:]
+        return result or options['format']
+
+    def _process_plantuml(self, options: CombinedOptions, body: str) -> str:
         '''Save PlantUML diagram body to .diag file, generate an image from it,
         and return the image ref.
 
@@ -77,7 +112,6 @@ class Preprocessor(BasePreprocessor):
 
         :returns: Image ref
         '''
-
         self.logger.debug(f'Processing PlantUML diagram, options: {options}, body: {body}')
 
         body_hash = md5(f'{body}'.encode())
@@ -87,20 +121,18 @@ class Preprocessor(BasePreprocessor):
 
         self.logger.debug(f'Diagram definition file path: {diagram_src_path}')
 
-        params = self.options.get('params', {})
+        # params = self.options.get('params', {})
 
-        diagram_format = {**params, **options}.get('format', 'png')
+        diagram_format = self._get_diagram_format(options)
 
         diagram_path = diagram_src_path.with_suffix(f'.{diagram_format}')
 
         self.logger.debug(f'Diagram image path: {diagram_path}')
 
-        img_ref = f'![{options.get("caption", "")}]({diagram_path.absolute().as_posix()})'
-
         if diagram_path.exists():
             self.logger.debug('Diagram image found in cache')
 
-            return img_ref
+            return self._get_result(diagram_path, options)
 
         diagram_src_path.parent.mkdir(parents=True, exist_ok=True)
 
@@ -134,7 +166,7 @@ class Preprocessor(BasePreprocessor):
                 diagram_path.rename(error_diagram_path)
 
                 error_message = (
-                    f'Processing of PlantUML diagram {diagram_src_path} failed: ' +
+                    f'Processing of PlantUML diagram {diagram_src_path} failed: '
                     f'{exception.output.decode()}'
                 )
 
@@ -145,7 +177,7 @@ class Preprocessor(BasePreprocessor):
             else:
                 raise RuntimeError(f'Failed: {exception.output.decode()}')
 
-        return img_ref
+        return self._get_result(diagram_path, options)
 
     def process_plantuml(self, content: str) -> str:
         '''Find diagram definitions and replace them with image refs.
@@ -158,14 +190,21 @@ class Preprocessor(BasePreprocessor):
         '''
 
         raw_pattern = re.compile(
-            r'^(\s*)(@startuml.+?@enduml)' +
+            r'^(\s*)(@startuml.+?@enduml)'
             r'|(?<=\n)(\s*)(@startuml.+?@enduml)',
             flags=re.DOTALL
         )
 
         def _sub(diagram) -> str:
+            options = CombinedOptions(
+                {
+                    'config': self.options,
+                    'tag': self.get_options(diagram.group('options'))
+                },
+                priority='tag'
+            )
             return self._process_plantuml(
-                self.get_options(diagram.group('options')),
+                options,
                 diagram.group('body')
             )
 
@@ -179,11 +218,12 @@ class Preprocessor(BasePreprocessor):
             diagram_groups = diagram.groups()
             spaces = diagram_groups[0] or diagram_groups[2]
             body = diagram_groups[1] or diagram_groups[3]
-            return spaces + self._process_plantuml({}, body)
+            return spaces + self._process_plantuml(self.options, body)
 
         # Process tags
         processed = self.pattern.sub(_sub, content)
-
+        with open('tmp.md', 'w') as f:
+            f.write(processed)
         # Process raw diagrams
         if self.options['parse_raw']:
             processed = raw_pattern.sub(_sub_raw, processed)